divide debug and internal api
