+++ ethereum/rpc/apis.go
@@ -78,7 +78,13 @@ func GetRPCAPIs(ctx *server.Context, clientCtx client.Context, tmWSClient *rpccl
 		{
 			Namespace: DebugNamespace,
 			Version:   apiVersion,
-			Service:   debug.NewDebugAPI(ctx),
+			Service:   debug.NewInternalAPI(ctx),
+			Public:    true,
+		},
+		{
+			Namespace: DebugNamespace,
+			Version:   apiVersion,
+			Service:   debug.NewDebugAPI(),
 			Public:    true,
 		},
 	}
